renamed a bunch of files, made the train have cars, and now the cars are having seizures.
